% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_plot.R
\name{basic_plot}
\alias{basic_plot}
\title{Provide the basic statistic graph for every variable in the dataframe}
\usage{
basic_plot(data)
}
\arguments{
\item{data}{target dataframe}
}
\value{

}
\description{
Provide the basic statistic graph for every variable in the dataframe
}
\examples{
basic_plot(dataframe1)
}
