% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealing_NAs.R
\name{dealing_NAs}
\alias{dealing_NAs}
\title{Dealing with NAs in the dataframe}
\usage{
dealing_NAs(data, method = "Default")
}
\arguments{
\item{data}{original dataframe}

\item{method}{"Default","delete": delete all observation with NAs;
"mean": impute the missing value by the mean of the column;
"knn": impute the missing value by knn method}
}
\value{

}
\description{
Dealing with NAs in the dataframe
}
\examples{
dealing_NAs(dataframe, method="knn")
}
