% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_series.R
\name{fred_series}
\alias{fred_series}
\title{Getdowm FRED dataset}
\usage{
fred_series(APIkey, series_id, observation_start, observation_end, frequency)
}
\arguments{
\item{APIkey}{personal FRED.api to get data down}

\item{series_id}{Abbreviation of the name of the time series you want}

\item{observation_start}{start date of your data}

\item{observation_end}{end date of your data}

\item{frequency}{frenquency of the time series: a = annual, m = monthly, w=weekly, d = daily}
}
\value{

}
\description{
Getdowm FRED dataset
}
\examples{
fred_series(APIkey="54b39d1f56f337fef917b0b1592874cc",series_id="CAUR",observation_start="1979-01-01",observation_end="2019-01-01",frequency="a")
}
