% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_select.R
\name{model_select}
\alias{model_select}
\title{Showing AICs and BICs for three models to select the best}
\usage{
model_select(model1, model2, model3)
}
\arguments{
\item{model1}{regression model 1}

\item{model2}{regression model 2}

\item{model3}{regression model 3}
}
\value{

}
\description{
Showing AICs and BICs for three models to select the best
}
\examples{
model_select(lm(y1~x1+x2),lm(y1~x1:x2),lm(y1~I(x1^2)+x2))
}
