% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ts.R
\name{summary_ts}
\alias{summary_ts}
\title{Summary of time series}
\usage{
summary_ts(
  timeseries,
  startdate,
  enddate,
  freq,
  plot = FALSE,
  decompose = FALSE
)
}
\arguments{
\item{timeseries}{a vector/column of a dataframe which is going to be changed into timeseries}

\item{startdate}{start date of the time series}

\item{enddate}{end date of the time series}

\item{freq}{frequency of the time series: 1=annual, 12=monthly, 4=quarterly}

\item{plot}{logical; present the graph of the time series or not}

\item{decompose}{logical; graph with the decomposition of the time series or not}
}
\value{

}
\description{
Summary of time series
}
\examples{
summary_ts(data1[,1],startdate=c(2003,5),enddate=c(2020,10),freq=12,plot=TRUE,decompose=TRUE)
}
